Splitted stream_reader from object detector
